Refactor content
