+++ llm_browser/src/database.py
@@ -1,1 +1,24 @@
 """MongoDB client setup, functions to interact with collections"""
+
+import os
+from urllib.parse import quote_plus
+
+from pymongo import MongoClient
+
+
+def get_mongodb_client():
+    """Establishes a MongoDB client
+
+    Returns
+    ---
+    Tuple of (MongoClient, database_name)
+    """
+
+    _USER = os.environ.get("_MONGO_UNAME")
+    _PASSWORD = quote_plus(os.environ.get("_MONGO_PWD"))
+    _HOST = os.environ.get("_MONGO_HOST")
+    _PORT = os.environ.get("_MONGO_PORT")
+
+    uri = f"mongodb://{_USER}:{_PASSWORD}@{_HOST}:{_PORT}/"
+
+    return MongoClient(uri)
